+++ backend/billing-module/src/main/java/com/vetclinic/billing/api/dto/InventoryItemResponse.java
@@ -20,7 +20,7 @@ public record InventoryItemResponse(
         String unit,
         Boolean active,
         String code,
-        Integer stockQuantity,
+        BigDecimal stockQuantity,
         Integer reorderPoint,
         String barcode,
         String supplierCode,
